// src/pages/TreatmentForm/types/TreatmentPlanFields.ts
export type TreatmentPlanFields = {
  id_medical_record?: string;
  restauracao_amalgama_c1?: string;
  restauracao_amalgama_c2_duas_faces?: string;
  restauracao_amalgama_c1_tres_faces?: string;
  restauracao_amalgama_c1_complexa?: string;
  restauracao_div?: string;
  restauracao_resina_c1?: string;
  restauracao_resina_c2_duas_faces?: string;
  restauracao_resina_c1_tres_faces?: string;
  restauracao_resina_c4?: string;
  restauracao_resina_c5?: string;
  restauracao_resina_complexa?: string;
  clareamento_vital?: string;
  clareamento_nao_vital?: string;
  exodontias?: string;
  selantes?: string;
  radiografias?: string;
  endodontias?: string;
  tratamento_periodontal?: string;
  protese_parcial_removivel?: string;
  protese_total?: string;
  protese_fixa?: string;
};